+++ Table.cs
@@ -6,14 +6,14 @@ namespace LuaExt;
 public sealed partial class Table
 {
 	[LuaMember("arrayCopy")]
-	public static LuaTable ArrayCopy(LuaTable table, int begin, int length) => table.ArrayCopy(begin, length);
+	public static LuaTable ArrayCopy(LuaTable table, double begin, double length) => table.ArrayCopy((int)begin, (int)length);
 
 	[LuaMember("getArrayCapacity")]
-	public static int GetArrayCapacity(LuaTable table) => table.GetArrayCapacity();
+	public static double GetArrayCapacity(LuaTable table) => table.GetArrayCapacity();
 
 	[LuaMember("getHashMapCount")]
-	public static int GetHashMapCount(LuaTable table) => table.HashMapCount;
+	public static double GetHashMapCount(LuaTable table) => table.HashMapCount;
 
 	[LuaMember("newTable")]
-	public static LuaTable NewTable(int arrayCapacity, int hashMapCapacity) => new(arrayCapacity, hashMapCapacity);
+	public static LuaTable NewTable(double arrayCapacity, double hashMapCapacity) => new((int)arrayCapacity, (int)hashMapCapacity);
 }
